add functionality for locations

+++ src/main/resources/startedData.sql
@@ -34,8 +34,8 @@ CREATE TABLE IF NOT EXISTS product
     description varchar(255),
     quantity    int,
     price       int,
-    catalog_id  bigint,
-    location_id bigint,
+    catalog_id  int,
+    location_id int,
     PRIMARY KEY (id)
 );
 --ALTER TABLE product ADD CONSTRAINT product_fk1 FOREIGN KEY (catalog_id) REFERENCES catalog (id) ON UPDATE RESTRICT ON DELETE RESTRICT;
